Extract 'styles/colors' module
