+++ TemplateProject/app/styles/colors.tsx
@@ -1,0 +1,7 @@
+const colors = {
+  primary: "#ed3e44",
+  secondary: "#29292c",
+  background: "white",
+};
+
+export { colors };
